feat(refactor-find-inactive-workplaces): Separate inactive workplace logic from next template logic
